requirements.txt
asgiref==3.4.1
bitstring==3.1.9
black==21.12b0
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
ifaddr==0.1.7
itsdangerous==2.0.1
Jinja2==3.0.3
lifxlan==1.2.7
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-dotenv==0.19.2
tomli==1.2.3
typing_extensions==4.0.1
Werkzeug==2.0.2
